Detect processes in view's initialProcessing container
Pipeline now include processes that create new views, by extracting them from initialProcessing container. However, this container belongs to the newly created view, not the target view of the process, and I couldn't find a way to deduct it. For now, these processes will be added to the pipeline with the target "UNDEFINED" and a TODO comment reminding they should be replaced with the correct target view. 
Minor changes on the output formatting.